requirements.txt
appdirs==1.4.3
attrs==19.3.0
click==7.1.2
distlib==0.3.0
entrypoints==0.3
filelock==3.0.12
flake8==3.7.9
Flask==1.1.2
Flask-SQLAlchemy==2.4.1
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.5.2
mccabe==0.6.1
more-itertools==8.2.0
mypy==0.770
mypy-extensions==0.4.3
packaging==20.3
pg8000==1.15.2
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.8.1
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.7
pytest==5.4.1
python-dotenv==0.13.0
scramp==1.1.1
six==1.14.0
SQLAlchemy==1.3.16
SQLAlchemy-Utils==0.36.4
testing.common.database==2.0.3
testing.postgresql==1.3.0
toml==0.10.0
tox==3.14.6
typed-ast==1.4.1
typing-extensions==3.7.4.2
virtualenv==20.0.18
wcwidth==0.1.9
Werkzeug==1.0.1
zipp==3.1.0
